refactor: split validation.js multi-domain Zod barrel into per-domain files (#1831)
validation.js (1,897 lines) mixed 15+ unrelated domains. Extract agent,
cos, media, and pipeline schema groups into agentValidation/cosValidation/
mediaValidation/pipelineValidation (continuing the #1151 split precedent),
leaving validation.js (~550 lines) with the app/process/provider/social/
backup/sharing core. validation.js re-exports the new files flat so every
existing deep import keeps working; the barrel adds namespace exports.
